feat!: upgrade system (part 1 of N) (#582)
## Breaking changes

- Changes in `ReplayRecord` stored format (new CFs: `ProtocolVersion`
and `ForcePreimages`)
- Changes in replay wire format (new `ReplayWireFormatV5` to work with
new replay fields)


## What does this PR support?

- Upgrade transactions are fetched from the contract and assembled
- Upgrade transactions are executed by the server once upgrade timestamp
is reached.
- Upgrade transaction will be the only tx in the block, and block with
upgrade transaction will be the only one in the batch.
- Batch with the new protocol version won't be committed until the
protocol upgrade is finalized on L1.
- Easy-to-implement integration tests for different upgrade flows
(including EN sync).

## What's NOT implemented?

These things are planned, but not included in this PR to keep the scope
reviewable.

- Matching protocol version to execution version (e.g. switching to a
new multivm version based on protocol upgrade).
- VKs switch (e.g. fetching new VKs from the protocol upgrade).
- Persistent state for partially fetched upgrade data.
- New multiVM system / artifact registry.

## Caveats

- ⚠️ We introduce breaking changes to both wire format and storage, we
need to think on how to roll out this feature.
- ⚠️ Protocol version for genesis is not exposed for ENs right now, so
they might not know which version should be used "initially". It's
unknown right now if it's a problem, since upgrade tx will be executed
in the first block.
- `BytecodesSupplier` is not ready for ZKsync OS, so the implementation
assumes that all the bytecodes will be deployed to L2 before the
upgrade. If this step won't be done, upgrade transaction will fail as
there is no way to fetch preimages from L1.
- Upgrades are expected to be sequential, e.g. node is not guaranteed to
correctly process multiple upgrades if somehow the previous upgrade
event was missing (e.g. if it happened a lot of time ago).
- Upgrades are expected to be executed in a reasonable time after it was
added to CTM (e.g. if upgrade was added to CTM, but executed after a few
months, we won't be able to find upgrade data).
- The way upgrade transactions are processed by `BlockContextProvider`
is not ideal, it exposes extra interfaces to the best txs stream, which
makes it a leaky abstraction. This problem will be solved once we have
persistent state for upgrade data.

## Implementation overview

### Changes to the existing components

- `BlockContextProvider` is reworked to receive upgrade transactions
from a channel. Old system with an `Option` for upgrade tx was removed.
Protocol version is now included in the produced command, and also
became a part of the replay record. Same for forced preimages (bytecodes
that must be inserted to state before executing the block, aka factory
deps).
- Genesis upgrade is processed in the same way as any other upgrade
transaction.
- `PersistentPreimages` no longer need genesis to be initialized; it
receives forced preimages from the upgrade transaction.
- `block_executor` will process upgrade transactions and seal the block
immediately after it's processed.
- `batch_builder` will seal the batch if it receives a block with
upgrade transaction (on both ends -- it will seal the previous batch
before starting a new one with upgrade tx, and one with upgrade tx will
be sealed with just 1 block)

### New components

- `UpgradeTxWatcher` - part of `l1_watcher`, looks for the upgrade txs
and sends them to the stream
- `UpgradeGatekeeper` - part of `l1_sender`, waits until protocol on
diamond proxy matches protocol version of the batch

## Testing

This PR introduces a way to (somewhat) declaratively do upgrade tests --
you can build the upgrade yourself, including setting a version, adding
force deployed contracts, etc, and then execute. `DefaultUpgrade` is
assumed but not forced. More functionality is planned to be added in the
future prs (e.g. setting VKs).
For now, 2 tests are added:
- "simple" upgrade (no force deployments)
- "normal" upgrade (with force deployments).

The latter also has a check for ENs to be able to sync.