feat: Read force deploys from a file (#612)
Since we do not yet support reading bytecodes using `BytecodeSupplier`
for zksync os, to temporarily enable upgrades, we will do the following:
- Each time there is a protocol upgrade tx, we just read the hardcoded
JSON file and include all the bytecodes from there to get the list of
bytecode hashes and their preimages.
- The correctness of the file is ensured by the person that prepares the
upgrade (most likely me). Hopefully this hack will not last for long.
- The file is included using `include_str!` so it is included into the
binary and so the binary can be shipped as-is just as before

PS: the file *will* change in the subsequent PRs/commits until the code
of contracts is frozen.

---------

Co-authored-by: Igor Aleksanov <popzxc@yandex.ru>